+++ cypress/integration/sharedSteps.js
@@ -125,3 +125,14 @@ And('you select the MNCH PNC program', () => {
         .click();
 });
 
+Then(/^the user clicks the element containing the text: (.*)$/, (text) => {
+    cy.contains(text).click();
+});
+
+Then(/^the current url is (.*)$/, (url) => {
+    cy.url().should('eq', `${Cypress.config().baseUrl}${url}`);
+});
+
+Then(/^the user ?(.*) see the following text: (.*)$/, (not, message) =>
+    cy.contains(message).should(not ? 'not.exist' : 'exist'),
+);