+++ src/core_modules/capture-core/components/Pages/EnrollmentEditEvent/EnrollmentEditEventPage.constants.js
@@ -1,0 +1,12 @@
+// @flow
+
+export const pageMode = Object.freeze({
+    EDIT: 'editEvent',
+    VIEW: 'viewEvent',
+});
+
+export const pageStatuses = {
+    DEFAULT: 'DEFAULT',
+    MISSING_DATA: 'MISSING_DATA',
+    WITHOUT_ORG_UNIT_SELECTED: 'WITHOUT_ORG_UNIT_SELECTED',
+};