+++ src/core_modules/capture-core/components/Pages/EnrollmentEditEvent/EnrollmentEditEventPage.types.js
@@ -1,6 +1,6 @@
 // @flow
 import type { ProgramStage } from '../../../metaData';
-import type { WidgetEffects, HideWidgets } from '../Enrollment/EnrollmentPageDefault/EnrollmentPageDefault.types';
+import type { WidgetEffects, HideWidgets } from '../common/EnrollmentOverviewDomain';
 
 export type PlainProps = {|
     programStage: ?ProgramStage,
@@ -10,7 +10,13 @@ export type PlainProps = {|
     enrollmentId: string,
     programId: string,
     mode: string,
+    orgUnitId: string,
+    trackedEntityName: string,
+    teiDisplayName: string,
+    eventDate?: string,
+    enrollmentsAsOptions: Array<Object>,
     onDelete: () => void,
     onGoBack: () => void,
+    pageStatus: string,
     ...CssClasses,
 |};