+++ src/core_modules/capture-core/components/ScopeSelector/ScopeSelector.types.js
@@ -8,7 +8,7 @@ export type OwnProps = $ReadOnly<{|
   pageToPush?: string,
   selectedOrgUnitId: string,
   selectedProgramId: string,
-  onSetProgramId: (id: string) => void,
+  onSetProgramId?: (id: string) => void,
   onResetProgramId: () => void,
   onSetOrgUnit: (id: string, orgUnit: Object) => void,
   onResetOrgUnitId: () => void,
@@ -20,14 +20,9 @@ export type PropsFromRedux = $ReadOnly<{|
 |}>
 
 export type DispatchersFromRedux = $ReadOnly<{|
-  onOpenNewEventPage: () => void,
-  onOpenNewRegistrationPageWithoutProgramId: Function,
-  onOpenSearchPage: () => void,
-  onOpenSearchPageWithoutProgramId: () => void,
   onSetCategoryOption: (categoryId: string, categoryOptionId: string) => void,
   onResetCategoryOption: (categoryId: string) => void,
   onResetAllCategoryOptions: () => void,
-  onStartAgain: () => void,
 |}>
 
 
@@ -42,10 +37,8 @@ export type Props = {|
 
 
 export type State = {|
-  openStartAgainWarning: boolean;
   openOrgUnitWarning: boolean;
   openProgramWarning: ?Object;
   openCatComboWarning: boolean;
   categoryIdToReset: string;
-  openNewEventWarning: boolean;
 |};