+++ src/core_modules/capture-core/components/ScopeSelector/hooks/index.js
@@ -6,6 +6,8 @@ export { useResetProgramId } from './useResetProgramId';
 export { useResetOrgUnitId } from './useResetOrgUnitId';
 export { useResetTeiId } from './useResetTeiId';
 export { useResetEnrollmentId } from './useResetEnrollmentId';
+export { useResetStageId } from './useResetStageId';
+export { useResetEventId } from './useResetEventId';
 export { useReset } from './useReset';
 
 export { useOrganizationUnit } from './useOrganizationUnit';