+++ src/core_modules/capture-core/components/WidgetStagesAndEvents/types/common.types.js
@@ -18,7 +18,8 @@ export type Stage = {
     description: string,
     icon?: Icon,
     dataElements: Array<StageDataElement>,
-    hideDueDate?: boolean
+    hideDueDate?: boolean,
+    repeatable?: boolean
 }
 
 export type StageCommonProps = {|