+++ src/epics/trackerCapture.epics.js
@@ -12,6 +12,9 @@ import {
     saveNewEventLocationChangeEpic,
     saveNewEventAddAnotherEpic,
     saveNewEventAddAnotherFailedEpic,
+    saveNewEventStageEpic,
+    saveNewEventStageFailedEpic,
+    saveNewEventInStageLocationChangeEpic,
     addNoteForNewSingleEventEpic,
     openRelationshipForNewSingleEventEpic,
     addRelationshipForNewSingleEventEpic,
@@ -200,14 +203,11 @@ import {
     openEnrollmentPageEpic,
 } from '../core_modules/capture-core/components/Pages/Enrollment/EnrollmentPage.epics';
 import {
-    cancelAddEventEpic,
-    cancelAddEventLocationChangeEpic,
-    runRulesOnUpdateDataEntryFieldForAddEventEpic,
-    runRulesOnUpdateFieldForAddEventEpic,
-    saveAddEventEpic,
-    saveAddEventLocationChangeEpic,
-    openAddEventInDataEntryEpic,
-} from '../core_modules/capture-core/components/WidgetAddEvent/NewEnrollmentEvent';
+    runRulesOnUpdateDataEntryFieldForNewEnrollmentEventEpic,
+    runRulesOnUpdateFieldForNewEnrollmentEventEpic,
+    saveNewEnrollmentEventEpic,
+    addNoteForNewEnrollmentEventEpic,
+} from '../core_modules/capture-core/components/WidgetEnrollmentEventNew';
 
 export const epics = combineEpics(
     resetProgramAfterSettingOrgUnitIfApplicableEpic,
@@ -330,11 +330,11 @@ export const epics = combineEpics(
     resetEnrollmentSelectionEpic,
     openEnrollmentPageEpic,
     checkForDuplicateEpic,
-    cancelAddEventEpic,
-    cancelAddEventLocationChangeEpic,
-    runRulesOnUpdateDataEntryFieldForAddEventEpic,
-    runRulesOnUpdateFieldForAddEventEpic,
-    saveAddEventEpic,
-    saveAddEventLocationChangeEpic,
-    openAddEventInDataEntryEpic,
+    saveNewEventStageEpic,
+    saveNewEventStageFailedEpic,
+    saveNewEventInStageLocationChangeEpic,
+    runRulesOnUpdateDataEntryFieldForNewEnrollmentEventEpic,
+    runRulesOnUpdateFieldForNewEnrollmentEventEpic,
+    saveNewEnrollmentEventEpic,
+    addNoteForNewEnrollmentEventEpic,
 );